+++ radix-engine/src/engine/transaction_receipt.rs
@@ -15,7 +15,7 @@ use crate::engine::RuntimeError;
 /// Represents a transaction receipt.
 pub struct Receipt {
     pub commit_receipt: Option<CommitReceipt>,
-    pub validated_instructions: Vec<ExecutableInstruction>,
+    pub instructions: Vec<ExecutableInstruction>,
     pub result: Result<(), RuntimeError>,
     pub outputs: Vec<ScryptoValue>,
     pub logs: Vec<(Level, String)>,
@@ -58,11 +58,11 @@ impl fmt::Debug for Receipt {
         )?;
 
         write!(f, "\n{}", "Instructions:".bold().green())?;
-        for (i, inst) in self.validated_instructions.iter().enumerate() {
+        for (i, inst) in self.instructions.iter().enumerate() {
             write!(
                 f,
                 "\n{} {}",
-                prefix!(i, self.validated_instructions),
+                prefix!(i, self.instructions),
                 match inst {
                     ExecutableInstruction::PublishPackage { .. } =>
                         "PublishPackage {..}".to_owned(),
