+++ radix-engine/tests/account.rs
@@ -56,7 +56,9 @@ fn cannot_withdraw_from_other_account() {
         .withdraw_from_account(RADIX_TOKEN, account)
         .call_method_with_all_resources(other_account, "deposit_batch")
         .build();
-    let signer_public_keys = vec![other_public_key]; // Act
+    let signer_public_keys = vec![other_public_key];
+
+    // Act
     let receipt = test_runner.execute_manifest(manifest, signer_public_keys);
 
     // Assert
@@ -80,7 +82,9 @@ fn account_to_bucket_to_account() {
                 .0
         })
         .build();
-    let signer_public_keys = vec![public_key]; // Act
+    let signer_public_keys = vec![public_key];
+
+    // Act
     let receipt = test_runner.execute_manifest(manifest, signer_public_keys);
 
     // Assert
@@ -95,7 +99,9 @@ fn test_account_balance() {
     let manifest = ManifestBuilder::new()
         .call_method(account, call_data![balance(RADIX_TOKEN)])
         .build();
-    let signer_public_keys = vec![public_key]; // Act
+    let signer_public_keys = vec![public_key];
+
+    // Act
     let receipt = test_runner.execute_manifest(manifest, signer_public_keys);
 
     // Assert
