+++ radix-engine/tests/authorization_dynamic.rs
@@ -1,4 +1,4 @@
-#[rustfmt::_ip]
+#[rustfmt::skip]
 pub mod test_runner;
 
 use crate::test_runner::TestRunner;
