+++ transaction/src/model/mod.rs
@@ -1,5 +1,7 @@
 mod executable;
+mod instruction;
 mod transaction;
 
 pub use self::transaction::*;
 pub use executable::*;
+pub use instruction::*;
